requirements.txt
fastapi
uvicorn
pydantic
requests
pytest-asyncio
httpx